+++ packages/cinterop/src/darwin/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -20,6 +20,7 @@ package io.realm.internal.interop
 
 import io.realm.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
 import io.realm.internal.interop.sync.AuthProvider
+import io.realm.internal.interop.sync.CoreUserState
 import io.realm.internal.interop.sync.MetadataMode
 import io.realm.internal.interop.sync.NetworkTransport
 import io.realm.mongodb.AppException
@@ -84,6 +85,7 @@ import realm_wrapper.realm_scheduler_t
 import realm_wrapper.realm_string_t
 import realm_wrapper.realm_sync_client_metadata_mode
 import realm_wrapper.realm_t
+import realm_wrapper.realm_user_t
 import realm_wrapper.realm_value_t
 import realm_wrapper.realm_value_type
 import realm_wrapper.realm_version_id_t
@@ -895,16 +897,21 @@ actual object RealmInterop {
         return CPointerWrapper(realm_wrapper.realm_app_get(appConfig.cptr(), syncClientConfig.cptr()))
     }
 
+    actual fun realm_app_get_current_user(app: NativePointer): NativePointer? {
+        val currentUserPtr: CPointer<realm_user_t>? = realm_wrapper.realm_app_get_current_user(app.cptr())
+        return nativePointerOrNull(currentUserPtr)
+    }
+
     actual fun realm_app_log_in_with_credentials(
         app: NativePointer,
         credentials: NativePointer,
-        callback: CinteropCallback
+        callback: AppCallback<NativePointer>
     ) {
         realm_wrapper.realm_app_log_in_with_credentials(
             app.cptr(),
             credentials.cptr(),
             staticCFunction { userdata, user, error: CPointer<realm_app_error_t>? ->
-                val userDataCallback = safeUserData<CinteropCallback>(userdata)
+                val userDataCallback = safeUserData<AppCallback<NativePointer>>(userdata)
                 if (error == null) {
                     // Remember to clone user object or else it will be invalidated right after we leave this callback
                     val clonedUser = realm_clone(user)
@@ -917,61 +924,48 @@ actual object RealmInterop {
                 }
             },
             StableRef.create(callback).asCPointer(),
-            staticCFunction { userdata -> disposeUserData<Callback>(userdata) }
+            staticCFunction { userdata -> disposeUserData<AppCallback<NativePointer>>(userdata) }
         )
     }
 
-    private val newRequestLambda = staticCFunction<COpaquePointer?,
-        CValue<realm_http_request_t>,
-        COpaquePointer?,
-        Unit>
-    { userdata, request, requestContext ->
-        safeUserData<NetworkTransport>(userdata).let { networkTransport ->
-            request.useContents { // this : realm_http_request_t ->
-                val headerMap = mutableMapOf<String, String>()
-                for (i in 0 until num_headers.toInt()) {
-                    headers?.get(i)?.let { header ->
-                        headerMap[header.name!!.toKString()] = header.value!!.toKString()
-                    } ?: error("Header at index $i within range ${num_headers.toInt()} should not be null")
-                }
-
-                val response = networkTransport.sendRequest(
-                    method = when (method) {
-                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_GET -> NetworkTransport.GET
-                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_POST -> NetworkTransport.POST
-                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_PATCH -> NetworkTransport.PATCH
-                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_PUT -> NetworkTransport.PUT
-                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_DELETE -> NetworkTransport.DELETE
-                    },
-                    url = url!!.toKString(),
-                    headers = headerMap,
-                    body = body!!.toKString()
-                )
-
-                memScoped {
-                    val headersSize = response.headers.entries.size
-                    val cResponseHeaders = allocArray<realm_http_header_t>(headersSize)
-
-                    response.headers.entries.forEachIndexed { i, entry ->
-                        cResponseHeaders[i].let { header ->
-                            header.name = entry.key.cstr.getPointer(memScope)
-                            header.value = entry.value.cstr.getPointer(memScope)
-                        }
+    actual fun realm_app_log_out(
+        app: NativePointer,
+        user: NativePointer,
+        callback: AppCallback<Unit>
+    ) {
+        realm_wrapper.realm_app_log_out(
+            app.cptr(),
+            user.cptr(),
+            staticCFunction { userdata, error: CPointer<realm_app_error_t>? ->
+                val userDataCallback = safeUserData<AppCallback<Unit>>(userdata)
+                if (error == null) {
+                    userDataCallback.onSuccess(Unit)
+                } else {
+                    val message = with(error.pointed) {
+                        "${message?.toKString()} [error_category=${error_category.value}, error_code=$error_code, link_to_server_logs=$link_to_server_logs]"
                     }
-
-                    val cResponse = alloc<realm_http_response_t> {
-                        body = response.body.cstr.getPointer(memScope)
-                        body_size = response.body.cstr.getBytes().size.toULong()
-                        custom_status_code = response.customResponseCode
-                        status_code = response.httpResponseCode
-                        num_headers = response.headers.entries.size.toULong()
-                        headers = cResponseHeaders
-                    }
-
-                    realm_wrapper.realm_http_transport_complete_request(requestContext, cResponse.ptr)
+                    userDataCallback.onError(AppException(message))
                 }
-            }
-        }
+            },
+            StableRef.create(callback).asCPointer(),
+            staticCFunction { userdata -> disposeUserData<AppCallback<NativePointer>>(userdata) }
+        )
+    }
+
+    actual fun realm_user_get_identity(user: NativePointer): String {
+        return realm_wrapper.realm_user_get_identity(user.cptr())!!.toKString()
+    }
+
+    actual fun realm_user_is_logged_in(user: NativePointer): Boolean {
+        return realm_wrapper.realm_user_is_logged_in(user.cptr())
+    }
+
+    actual fun realm_user_log_out(user: NativePointer) {
+        realm_wrapper.realm_user_log_out(user.cptr())
+    }
+
+    actual fun realm_user_get_state(user: NativePointer): CoreUserState {
+        return CoreUserState.of(realm_wrapper.realm_user_get_state(user.cptr()))
     }
 
     actual fun realm_sync_client_config_new(): NativePointer {
@@ -1137,6 +1131,14 @@ actual object RealmInterop {
         realm_wrapper.realm_config_set_sync_config(realmConfiguration.cptr(), syncConfiguration.cptr())
     }
 
+    private fun nativePointerOrNull(ptr: CPointer<*>?, managed: Boolean = true): NativePointer? {
+        return if (ptr != null) {
+            CPointerWrapper(ptr, managed)
+        } else {
+            null
+        }
+    }
+
     private fun MemScope.classInfo(realm: NativePointer, table: String): realm_class_info_t {
         val found = alloc<BooleanVar>()
         val classInfo = alloc<realm_class_info_t>()
@@ -1238,6 +1240,59 @@ actual object RealmInterop {
                 }
             }
         )
+    }
+
+    private val newRequestLambda = staticCFunction<COpaquePointer?,
+        CValue<realm_http_request_t>,
+        COpaquePointer?,
+        Unit>
+    { userdata, request, requestContext ->
+        safeUserData<NetworkTransport>(userdata).let { networkTransport ->
+            request.useContents { // this : realm_http_request_t ->
+                val headerMap = mutableMapOf<String, String>()
+                for (i in 0 until num_headers.toInt()) {
+                    headers?.get(i)?.let { header ->
+                        headerMap[header.name!!.toKString()] = header.value!!.toKString()
+                    } ?: error("Header at index $i within range ${num_headers.toInt()} should not be null")
+                }
+
+                val response = networkTransport.sendRequest(
+                    method = when (method) {
+                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_GET -> NetworkTransport.GET
+                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_POST -> NetworkTransport.POST
+                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_PATCH -> NetworkTransport.PATCH
+                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_PUT -> NetworkTransport.PUT
+                        realm_http_request_method.RLM_HTTP_REQUEST_METHOD_DELETE -> NetworkTransport.DELETE
+                    },
+                    url = url!!.toKString(),
+                    headers = headerMap,
+                    body = body!!.toKString()
+                )
+
+                memScoped {
+                    val headersSize = response.headers.entries.size
+                    val cResponseHeaders = allocArray<realm_http_header_t>(headersSize)
+
+                    response.headers.entries.forEachIndexed { i, entry ->
+                        cResponseHeaders[i].let { header ->
+                            header.name = entry.key.cstr.getPointer(memScope)
+                            header.value = entry.value.cstr.getPointer(memScope)
+                        }
+                    }
+
+                    val cResponse = alloc<realm_http_response_t> {
+                        body = response.body.cstr.getPointer(memScope)
+                        body_size = response.body.cstr.getBytes().size.toULong()
+                        custom_status_code = response.customResponseCode
+                        status_code = response.httpResponseCode
+                        num_headers = response.headers.entries.size.toULong()
+                        headers = cResponseHeaders
+                    }
+
+                    realm_wrapper.realm_http_transport_complete_request(requestContext, cResponse.ptr)
+                }
+            }
+        }
     }
 
     data class CoreCallback(