+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/App.kt
@@ -61,6 +61,20 @@ interface App {
     val emailPasswordAuth: EmailPasswordAuth
 
     /**
+     * Returns the current user that is logged in and still valid.
+     *
+     * A user is invalidated when he/she logs out or the user's refresh token expires or is revoked.
+     *
+     * If two or more users are logged in, it is the last valid user that is returned by this
+     * method.
+     *
+     * @return current [User] that has logged in and is still valid or `null` if no user is logged
+     * in or the user has expired.
+     */
+    // FIXME convert to property since it is state and not functionality?
+    fun currentUser(): User?
+
+    /**
      * Log in as a user with the given credentials associated with an authentication provider.
      *
      * @param credentials the credentials representing the type of login.